requirements.txt
PyYAML
requests
pandas
pytest
